requirements.txt
bcrypt
flask
markupsafe
mysql-connector-python